+++ InnerEye/ML/utils/ml_util.py
@@ -163,18 +163,6 @@ def set_random_seed(random_seed: int, caller_name: Optional[str] = None) -> None
     logging.debug(f"{prefix}Random seed set to: {random_seed}")
 
 
-# noinspection PyUnresolvedReferences,PyTypeHints
-def make_pytorch_reproducible() -> None:
-    """
-    Sets pytorch to a state such that 2 independent training runs do really give the same results.
-    """
-    # These two settings come from https://pytorch.org/docs/stable/notes/randomness.html
-    # Caveat: small increase in training time. For classification models, training time increased from
-    # 22:25min / 22:35min to 22:45min / 22:50min
-    torch.backends.cudnn.deterministic = True  # type: ignore
-    torch.backends.cudnn.benchmark = False  # type: ignore
-
-
 def is_test_from_execution_mode(execution_mode: ModelExecutionMode) -> bool:
     """
     Returns a boolean by checking the execution type. The output is used to determine the properties